+++ familyhelper-webapi-node/src/main/java/com/dwarfeng/familyhelper/webapi/node/configuration/settingrepo/BeanTransformerConfiguration.java
@@ -1,0 +1,38 @@
+package com.dwarfeng.familyhelper.webapi.node.configuration.settingrepo;
+
+import com.dwarfeng.settingrepo.sdk.bean.entity.FastJsonFormatterSupport;
+import com.dwarfeng.settingrepo.sdk.bean.entity.FastJsonSettingCategory;
+import com.dwarfeng.settingrepo.sdk.bean.entity.FastJsonSettingNode;
+import com.dwarfeng.settingrepo.stack.bean.entity.FormatterSupport;
+import com.dwarfeng.settingrepo.stack.bean.entity.SettingCategory;
+import com.dwarfeng.settingrepo.stack.bean.entity.SettingNode;
+import com.dwarfeng.subgrade.impl.bean.DozerBeanTransformer;
+import com.dwarfeng.subgrade.stack.bean.BeanTransformer;
+import org.dozer.Mapper;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration("settingrepo.beanTransformerConfiguration")
+public class BeanTransformerConfiguration {
+
+    private final Mapper mapper;
+
+    public BeanTransformerConfiguration(Mapper mapper) {
+        this.mapper = mapper;
+    }
+
+    @Bean("settingrepo.settingNodeBeanTransformer")
+    public BeanTransformer<SettingNode, FastJsonSettingNode> settingNodeBeanTransformer() {
+        return new DozerBeanTransformer<>(SettingNode.class, FastJsonSettingNode.class, mapper);
+    }
+
+    @Bean("settingrepo.settingCategoryBeanTransformer")
+    public BeanTransformer<SettingCategory, FastJsonSettingCategory> settingCategoryBeanTransformer() {
+        return new DozerBeanTransformer<>(SettingCategory.class, FastJsonSettingCategory.class, mapper);
+    }
+
+    @Bean("settingrepo.formatterSupportBeanTransformer")
+    public BeanTransformer<FormatterSupport, FastJsonFormatterSupport> formatterSupportBeanTransformer() {
+        return new DozerBeanTransformer<>(FormatterSupport.class, FastJsonFormatterSupport.class, mapper);
+    }
+}