+++ familyhelper-webapi-node/src/main/java/com/dwarfeng/familyhelper/webapi/node/controller/v1/settingrepo/FormatterSupportController.java
@@ -1,0 +1,102 @@
+package com.dwarfeng.familyhelper.webapi.node.controller.v1.settingrepo;
+
+import com.dwarfeng.familyhelper.webapi.stack.service.settingrepo.FormatterSupportResponseService;
+import com.dwarfeng.settingrepo.sdk.bean.entity.FastJsonFormatterSupport;
+import com.dwarfeng.settingrepo.stack.bean.entity.FormatterSupport;
+import com.dwarfeng.subgrade.sdk.bean.dto.FastJsonResponseData;
+import com.dwarfeng.subgrade.sdk.bean.dto.JSFixedFastJsonPagedData;
+import com.dwarfeng.subgrade.sdk.bean.dto.PagingUtil;
+import com.dwarfeng.subgrade.sdk.bean.dto.ResponseDataUtil;
+import com.dwarfeng.subgrade.sdk.interceptor.analyse.BehaviorAnalyse;
+import com.dwarfeng.subgrade.sdk.interceptor.analyse.SkipRecord;
+import com.dwarfeng.subgrade.sdk.interceptor.login.LoginRequired;
+import com.dwarfeng.subgrade.stack.bean.BeanTransformer;
+import com.dwarfeng.subgrade.stack.bean.dto.PagedData;
+import com.dwarfeng.subgrade.stack.bean.dto.PagingInfo;
+import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
+import com.dwarfeng.subgrade.stack.exception.ServiceExceptionMapper;
+import org.springframework.web.bind.annotation.*;
+
+import javax.servlet.http.HttpServletRequest;
+
+/**
+ * 格式化器支持控制器。
+ *
+ * @author DwArFeng
+ * @since 1.0.6
+ */
+@RestController
+@RequestMapping("/api/v1/settingrepo")
+public class FormatterSupportController {
+
+    private final FormatterSupportResponseService service;
+    private final ServiceExceptionMapper sem;
+
+    private final BeanTransformer<FormatterSupport, FastJsonFormatterSupport> beanTransformer;
+
+    public FormatterSupportController(
+            FormatterSupportResponseService service, ServiceExceptionMapper sem,
+            BeanTransformer<FormatterSupport, FastJsonFormatterSupport> beanTransformer
+    ) {
+        this.service = service;
+        this.sem = sem;
+        this.beanTransformer = beanTransformer;
+    }
+
+    @GetMapping("/formatter-support/{id}/exists")
+    @BehaviorAnalyse
+    @LoginRequired
+    public FastJsonResponseData<Boolean> exists(HttpServletRequest request, @PathVariable("id") String id) {
+        try {
+            boolean exists = service.exists(new StringIdKey(id));
+            return FastJsonResponseData.of(ResponseDataUtil.good(exists));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(Boolean.class, e, sem));
+        }
+    }
+
+    @GetMapping("/formatter-support/{id}")
+    @BehaviorAnalyse
+    @LoginRequired
+    public FastJsonResponseData<FastJsonFormatterSupport> get(HttpServletRequest request, @PathVariable("id") String id) {
+        try {
+            FormatterSupport formatterSupport = service.get(new StringIdKey(id));
+            return FastJsonResponseData.of(ResponseDataUtil.good(FastJsonFormatterSupport.of(formatterSupport)));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(FastJsonFormatterSupport.class, e, sem));
+        }
+    }
+
+    @GetMapping("/formatter-support/all")
+    @BehaviorAnalyse
+    @SkipRecord
+    @LoginRequired
+    public FastJsonResponseData<JSFixedFastJsonPagedData<FastJsonFormatterSupport>> all(
+            HttpServletRequest request, @RequestParam("page") int page, @RequestParam("rows") int rows
+    ) {
+        try {
+            PagedData<FormatterSupport> all = service.all(new PagingInfo(page, rows));
+            PagedData<FastJsonFormatterSupport> transform = PagingUtil.transform(all, beanTransformer);
+            return FastJsonResponseData.of(ResponseDataUtil.good(JSFixedFastJsonPagedData.of(transform)));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(JSFixedFastJsonPagedData.class, e, sem));
+        }
+    }
+
+    @GetMapping("/formatter-support/id-like")
+    @BehaviorAnalyse
+    @SkipRecord
+    @LoginRequired
+    public FastJsonResponseData<JSFixedFastJsonPagedData<FastJsonFormatterSupport>> idLike(
+            HttpServletRequest request,
+            @RequestParam("pattern") String pattern, @RequestParam("page") int page, @RequestParam("rows") int rows
+    ) {
+        try {
+            PagedData<FormatterSupport> all = service.idLike(pattern, new PagingInfo(page, rows));
+            PagedData<FastJsonFormatterSupport> transform = PagingUtil.transform(all, beanTransformer);
+            return FastJsonResponseData.of(ResponseDataUtil.good(JSFixedFastJsonPagedData.of(transform)));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(JSFixedFastJsonPagedData.class, e, sem));
+        }
+    }
+}